setup.py: install_requires: add modules by hand
Do not read the install_requires value from requirements.txt file.

"It is not considered best practice to use install_requires to pin
dependencies to specific versions, or to specify sub-dependencies (i.e.
dependencies of your dependencies). This is overly-restrictive, and
prevents the user from gaining the benefit of dependency upgrades."

Lastly, it’s important to understand that install_requires is a listing
of “Abstract” requirements, i.e just names and version restrictions that
don’t determine where the dependencies will be fulfilled from (i.e. from
what index or source). The where (i.e. how they are to be made
“Concrete”) is to be determined at install time using pip options."

 -- https://packaging.python.org/discussions/install-requires-vs-requirements/