fix(worker.py): recover from incorrectly constructed Inquiry message
Sending an invalid `Inquiry` message, in some cases, would cause the
worker to leave a dangling socket while connected to the broker. This
would occur, for example, when a client sent an `Inquiry` message
without declaring the type of the inquiry in the `oneof` field.

"zmq.error.ZMQError: Operation cannot be accomplished in current state"

Gracefully report the error to the client, close the socket and
reconnect to the broker.